Create Ark Dynamic Config File
The DynamicConfig allows to change some server settings without the need of restarting the server. The file must be hosted on a web server (e.g.: IIS on Windows, Apache etc...) using the HTTP protocol (HTTPS is not supported) and the .ini file extension must be exposed as mime type as text/plain. It cannot be directly linked using a system path. Once -UseDynamicConfig and CustomDynamicConfigUrl options are set, the following settings are available to dynamically configure: